Handle pending Erc20 transactions
- Make timestamp types Int instead of TimeInterval

+++ Example/EthereumKit/Core/TransactionRecord.swift
@@ -7,7 +7,7 @@ struct TransactionRecord {
     let transactionIndex: Int
     let interTransactionIndex: Int
     let amount: Decimal
-    let timestamp: Double
+    let timestamp: Int
 
     let from: TransactionAddress
     let to: TransactionAddress
